1.17.0 — Float→BCMath migration packs 1-12 + indicators_synced_at skip-stamp
107 net (float) casts removed from kraitebot/core/src/ across the
order mappers, recovery, drift checking, exchange-information mappers,
billing wallet, position accessors, market-regime helpers,
SupportResistanceProximity, AssignBestTokens one-way direction filter,
and PriceVolatilityIndicator. Precision-sensitive arithmetic now routes
through Kraite\Core\Support\Math::* instead of IEEE-754 floats. The
remaining 91 casts are legitimate boundaries (display coercion,
log-using statistics, return-type contracts, downstream consumer
cascades, final API surfaces) and are classified KEEP.

Bug fix: ConcludeSymbolDirectionAtTimeframeJob's same_indicator_data
skip branch now stamps indicators_synced_at like the other successful
end-to-end branches. Without this, long-timeframe symbols (1d) sat
past the system-health watchdog's 90-minute threshold for nearly the
full day even when the conclude pipeline ran every cron tick. The
column is a liveness signal for the watchdog — every successful
end-to-end run must stamp it.

+++ src/Support/Recovery/AbstractPositionRecoverer.php
@@ -418,7 +418,9 @@ abstract class AbstractPositionRecoverer
         if ($tp !== null && Math::gt((string) $tp->price, '0') && Math::gt($currentBreakeven, '0')) {
             $diff = self::absDelta($currentBreakeven, (string) $tp->price);
             $profitPct = Math::mul(Math::div($diff, $currentBreakeven, 8), '100', 3);
-            $updates['profit_percentage'] = (float) $profitPct;
+            // DECIMAL columns accept the BCMath string directly —
+            // skipping the float cast keeps full precision.
+            $updates['profit_percentage'] = $profitPct;
 
             // Anchor #2: first_profit_price computed from the original
             // entry + the percentage. For LONG, TP is above entry; for
@@ -434,7 +436,7 @@ abstract class AbstractPositionRecoverer
 
         if ($sl !== null && Math::gt((string) $sl->price, '0') && Math::gt($currentBreakeven, '0')) {
             $diff = self::absDelta($currentBreakeven, (string) $sl->price);
-            $updates['stop_market_percentage'] = (float) Math::mul(Math::div($diff, $currentBreakeven, 8), '100', 2);
+            $updates['stop_market_percentage'] = Math::mul(Math::div($diff, $currentBreakeven, 8), '100', 2);
         }
 
         if ($updates !== []) {
